Update modal height to give bottom offset of 10%, just like the top
We can't simply set bottom to 10%, otherwise all modals will be that full size, no matter how short the content. Using a max-height of 100vh minus the top and bottom offset that we want, allows us to have the modal grow as large as it needs to be, but then not going further than the bottom of the viewport.
This change does have a side effect on the right and left variations of the modal, so that has been corrected by "resetting" the max-height value for those variations